==> R/00_data_prep.Rmd <==
---
title: "Dog-SEP"
subtitle: "Data preps"
pagetitle: "Dog-SEP: preps"
author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document: 
    self_contained: true
    highlight: pygments
    keep_md: no
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 4
    toc_float: yes
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ------------------------------------------------------------ --> 

```{r r-setup, include = FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")

library(pacman)
p_load(tidyverse, readxl, magrittr, scales,
       janitor, sjmisc, sjPlot, DT, gmodels,
       sf, tmap,
       caret)

tmap_mode("view") # makes map interactive
```

```{r knit-setup, include = FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, 
                      fig.width = 8, fig.height = 6, dpi = 300, 
                      out.width = "800px", out.height = "600px")

knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)

knitr::opts_knit$set(width = 75)

mu <- Hmisc::markupSpecs$html
```

<!-- ------------------------------------------------------------ --> 

# Neighbourhood boundaries

## BCC suburbs

```{r eval=FALSE}
# orig data missing???
SUB <- read_csv("data-raw/geo/suburb-and-adjoining-suburb-november-2019.zip") %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>% 
  select(suburb_name) %>% 
  rename(SSC_NAME16 = suburb_name) %>%  
  distinct() %>% 
  mutate(SSC_NAME16 = str_to_title(SSC_NAME16)) %>% 
  arrange(SSC_NAME16)

write_rds(SUB, "data/geo/clean/SUB.Rds")
```

```{r include=FALSE}
SUB <- read_rds("data/geo/SUB.Rds")

# glimpse(SUB)
```

This might very well include areas with no pops (and therefore no SEIFA), for instance:

```{r echo=FALSE}
SUB %>% 
  filter(str_detect(SSC_NAME16, 
                    regex("port", ignore_case = TRUE)))
```

## ABS

Names from Brisbane containing `(Brisbane - Qld)`, names from Qld containing `(Qld)` have to be cleaned to match BCC data.

`Mcdowall` is called `McDowall` and `Mount Coot-tha` is `Mount Coot-Tha` - these have been unified as well.

```{r eval=FALSE}
unzip("data-raw/geo/1270055003_ssc_2016_aust_shape.zip", 
      exdir = "data-raw/geo")

SSC <- st_read("data-raw/geo/1270055003_ssc_2016_aust_shape/SSC_2016_AUST.shp", 
               stringsAsFactors = FALSE) %>% 
  mutate(SSC_CODE16 = as.integer(SSC_CODE16)) %>% 
  select(-STE_NAME16, -STE_CODE16, -AREASQKM16) %>% 	
  st_transform(3112) %>% 	
  filter(!st_is_empty(geometry)) %>% 
  mutate(SSC_NAME16_orig = SSC_NAME16) %>%  
  mutate(SSC_NAME16 = str_remove(SSC_NAME16, 
                                 fixed(" (Brisbane - Qld)"))) %>% 
  mutate(SSC_NAME16 = str_remove(SSC_NAME16, 
                                 fixed(" (Qld)"))) %>% 
  mutate(SSC_NAME16 = ifelse(SSC_NAME16 == "McDowall",
                             "Mcdowall", SSC_NAME16)) %>% 
  mutate(SSC_NAME16 = ifelse(SSC_NAME16 == "Mount Coot-tha",
                             "Mount Coot-Tha", SSC_NAME16))

# SSC <- rmapshaper::ms_simplify(SSC, keep = 0.05, weighting = 0.7) # default settings

write_rds(SSC, "data/geo/SSC_2016_AUST.Rds")

unlink("data-raw/geo/1270055003_ssc_2016_aust_shape", recursive = TRUE)
```

```{r include=FALSE}
SSC <- read_rds("data/geo/SSC_2016_AUST.Rds")
```

```{r eval=FALSE, include=FALSE}
SUB %>% 
  left_join(SSC) %>% 
  select(-geometry) %>% 
  filter(SSC_NAME16 != SSC_NAME16_orig) 
```

`Stones Corner` doesn't exist in ABS but it does in BCC. It seems it's part of [`Greenslopes`](https://quickstats.censusdata.abs.gov.au/census_services/getproduct/census/2016/quickstat/SSC31254?opendocument).

```{r echo=FALSE}
SUB %>% 
  left_join(SSC) %>% 
  select(-geometry) %>% 
  filter(is.na(SSC_CODE16))
```

After clean the datasets were merged.  

```{r echo=FALSE}
SSC %<>% 
  right_join(SUB) %>%
  filter(SSC_NAME16 != "Stones Corner")
```

Full map

```{r echo=FALSE}
# SSC %>% 
#   plot(max.plot = 1)

qtm(SSC, fill = NULL, borders = "darkorchid4", 
    text ="SSC_NAME16", text.col = "darkorchid4")
```

```{r eval=FALSE, include=FALSE}
## QSC
# data not needed at the moment 
# kept as backup solution in case of problem with source above

unzip("data-raw/geo/QSC_Extracted_Data_20190816_111155192000-13108.zip")

LB <- st_read("data-raw/geo/QSC_Extracted_Data_20190816_111155192000-13108/Locality_Boundaries.shp", 
              stringsAsFactors = FALSE) %>% 
  select(-ADMINTYPEN, -CA_AREA_SQ, -SHAPE_Leng, -SHAPE_Area) %>% 
  mutate(LOC_CODE = as.integer(LOC_CODE)) %>% 
  filter(!st_is_empty(geometry)) %>%
  filter(LGA == "Brisbane City") %>%
  select(-LGA) %>%
  # st_drop_geometry() %>% 
  st_transform(3112)

# LB <- ms_simplify(LB, keep = 0.05, weighting = 0.7) # default settings

write_rds(LB, "data/geo/Locality_Boundaries.Rds")

plot(st_geometry(LB))

qtm(LB, fill = NULL, borders = "darkorchid4", 
    text ="LOCALITY", text.col = "darkorchid4")

test <- LB %>% 
  st_drop_geometry() %>% 
  left_join(SSC, by = c("LOCALITY" = "SSC_NAME16")) %>% 
  select(-geometry)

test <- LB %>% 
  st_drop_geometry() %>% 
  left_join(SUB %>% mutate(exist = 1), 
            by = c("LOCALITY" = "SSC_NAME16"))

# Three changes with capital letter problems
LB %<>% 
  mutate(LOCALITY = ifelse(LOCALITY == "MacGregor",
                           "Macgregor", LOCALITY)) %>%
  mutate(LOCALITY = ifelse(LOCALITY == "McDowall",
                           "Mcdowall", LOCALITY)) %>%
  mutate(LOCALITY = ifelse(LOCALITY == "Mount Coot-tha",
                           "Mount Coot-Tha", LOCALITY))

# Now one extra 'suburb' of Moreton Bay
LB %>% 
  st_drop_geometry() %>% 
  left_join(SUB %>% mutate(exist = 1), 
            by = c("LOCALITY" = "SSC_NAME16")) %>% 
  filter(is.na(exist)) %>% 
  select(LOCALITY)

# And one extra Stones Creek 
LB %>% 
  st_drop_geometry() %>% 
  filter(LOCALITY == "Stones Corner")

unlink("data-raw/geo/QSC_Extracted_Data_20190816_111155192000-13108", recursive = TRUE)
```

```{r include=FALSE}
rm(SUB)
```

<!-- ------------------------------------------------------------ --> 

# SEIFA

```{r message=FALSE, warning=FALSE}
SEIFA <- read_xls("data-raw/SEIFA/2033055001 - ssc indexes.xls", 
                  sheet = "Table 1", skip = 5, n_max = 13719, na = "-") %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>% 
  dplyr::rename(SSC_CODE16 = x1,
                SSC_NAME16 = x2,
                IRSD = score_3,
                IRSD_d = decile_4,
                IRSAD = score_5,
                IRSAD_d = decile_6,
                IER = score_7,
                IER_d = decile_8,
                IEO = score_9,
                IEO_d = decile_10,
                URP = x11,
                caution = x12) %>% 
  # deciles kept as integers here
  mutate(SSC_CODE16 = as.integer(SSC_CODE16),
         IRSD = as.integer(IRSD),
         IRSAD = as.integer(IRSAD),
         IER = as.integer(IER),
         IEO = as.integer(IEO),
         IRSD_d = as.integer(IRSD_d),
         IRSAD_d = as.integer(IRSAD_d),
         IER_d = as.integer(IER_d),
         IEO_d = as.integer(IEO_d),
         URP = as.integer(URP)
  ) %>% 
  mutate(caution = as.logical(ifelse(is.na(caution), "False", "True")))
```

```{r include=FALSE}
write_rds(SEIFA, "data/SEIFA/SEIFA.Rds")

SEIFA %<>% 
  select(-SSC_NAME16)

SSC %<>% 
  left_join(SEIFA) 

# SSC %>% 
#   st_drop_geometry() %>% 
#   glimpse()
```

## Coverage 

Comparing Australia & Brisbane, example of IRSD.

### Frequencies table {.tabset}

#### Australia 

```{r echo=FALSE}
frq(SEIFA$IRSD_d)
```

#### Brisbane  

```{r echo=FALSE}
frq(SSC$IRSD_d)
```

### Frequencies plot {.tabset}

#### Australia 

```{r echo=FALSE}
SEIFA %>% 
  filter(!is.na(IRSD_d)) %>% 
  ggplot(aes(x = as.factor(IRSD_d))) + 
  geom_bar(aes(y = (..count..)/sum(..count..))) + 
  scale_y_continuous(labels = scales::percent) +
  ylab("% of areas")
```

#### Brisbane  

```{r echo=FALSE}
SSC %>% 
  filter(!is.na(IRSD_d)) %>% 
  ggplot(aes(x = as.factor(IRSD_d))) + 
  geom_bar(aes(y = (..count..)/sum(..count..))) + 
  scale_y_continuous(labels = scales::percent) +
  ylab("% of areas")
```

### Range per decile {.tabset}

#### Australia 

```{r echo=FALSE}
SEIFA %>% 
  filter(!is.na(IRSD_d)) %>% 
  ggplot(aes(x = as.factor(IRSD_d), y = IRSD)) + 
  geom_boxplot(varwidth = TRUE) + 
  ylim(min(SEIFA$IRSD, na.rm = TRUE), max(SEIFA$IRSD, na.rm = TRUE))
```

#### Brisbane  

```{r echo=FALSE}
SSC %>% 
  filter(!is.na(IRSD_d)) %>% 
  ggplot(aes(x = as.factor(IRSD_d), y = IRSD)) + 
  geom_boxplot(varwidth = TRUE) + 
  ylim(min(SEIFA$IRSD, na.rm = TRUE), max(SEIFA$IRSD, na.rm = TRUE))
```

## Missing

Few areas with missing SEIFA

```{r echo=FALSE}
SSC %>% 
  st_drop_geometry() %>% 
  filter_at(vars(ends_with("_d")), 
            any_vars(is.na(.))) %>%   
  select(SSC_NAME16, ends_with("_d"), URP)
```

These were excluded.

```{r include=FALSE}
SSC %<>% 
  filter_at(vars(ends_with("_d")), all_vars(!is.na(.))) 
```

## Caution

Few cases with ABS flag `caution`.

```{r eval=FALSE, include=FALSE}
frq(SSC$caution)
```

```{r echo=FALSE}
SSC %>% 
  st_drop_geometry() %>% 
  filter(caution) %>% 
  select(-caution, -SSC_CODE16, -SSC_NAME16_orig)
```

Usually with very small pop numbers.

```{r eval=FALSE, include=FALSE}
SSC %>% 
  st_drop_geometry() %>% 
  group_by(caution) %>% 
  summarise(mean = mean(URP),
            min = min(URP),
            max = max(URP))
```

These remain included.

## 'Local' deciles

Original values of indices were used to calculate 'local deciles' using SSCs for Brisbane only

```{r}
SSC %<>% 
  mutate(IRSD_d_orig = IRSD_d,
         IRSD_d = ntile(IRSD, 10),
         IRSAD_d_orig = IRSAD_d,
         IRSAD_d = ntile(IRSAD, 10),
         IER_d_orig = IER_d,
         IER_d = ntile(IER, 10),
         IEO_d_orig = IEO_d,
         IEO_d = ntile(IEO, 10)) %>% 
  mutate(across(ends_with("_d"), factor)) %>% 
  mutate(across(ends_with("_d_orig"), factor)) %>% 
  relocate(IRSD_d_orig, .after = IRSD_d) %>% 
  relocate(IRSAD_d_orig, .after = IRSAD_d) %>% 
  relocate(IER_d_orig, .after = IER_d) %>% 
  relocate(IEO_d_orig, .after = IEO_d) 
```

Note: Deciles are represented as `factor` variable. Certain application could benefit from changing it to `ordered`?  

Agreement across specific indices :  

### `IRSD_d` 

```{r echo=FALSE}
# frq(SSC$IRSD_d_orig)
# frq(SSC$IRSD_d)

SSC %>% 
  st_drop_geometry() %>% 
  frq(IRSD_d_orig == IRSD_d)
```

```{r echo=FALSE}
plot_xtab(SSC$IRSD_d, SSC$IRSD_d_orig, 
          margin = "row", bar.pos = "stack",
          show.summary = TRUE, coord.flip = TRUE)

tabz <- table(SSC$IRSD_d, SSC$IRSD_d_orig)

assocplot(tabz,
          xlab = "IRSD_d", ylab = "IRSD_d_orig")

mosaicplot(tabz,
           xlab = "IRSD_d", ylab = "IRSD_d_orig")
```

```{r echo=FALSE}
tm_shape(SSC) +
  tm_polygons(col = c("IRSD_d", "IRSD_d_orig"),
              n = 10, palette = "div",
              id = "SSC_NAME16", 
              popup.vars = c("SSC_NAME16", "IRSD_d", "IRSD"))
```

### `IRSAD_d` 

```{r echo=FALSE}
# frq(SSC$IRSAD_d_orig)
# frq(SSC$IRSAD_d)

SSC %>% 
  st_drop_geometry() %>% 
  frq(IRSAD_d_orig == IRSAD_d)
```

```{r echo=FALSE}
plot_xtab(SSC$IRSAD_d, SSC$IRSAD_d_orig, 
          margin = "row", bar.pos = "stack",
          show.summary = TRUE, coord.flip = TRUE)

tabz <- table(SSC$IRSAD_d, SSC$IRSAD_d_orig)

assocplot(tabz,
          xlab = "IRSAD_d", ylab = "IRSAD_d_orig")

mosaicplot(tabz,
           xlab = "IRSAD_d", ylab = "IRSAD_d_orig")
```

```{r echo=FALSE}
tm_shape(SSC) +
  tm_polygons(col = c("IRSAD_d", "IRSAD_d_orig"),
              n = 10, palette = "div",
              id = "SSC_NAME16", 
              popup.vars = c("SSC_NAME16", "IRSAD_d", "IRSAD"))
```

### `IER_d` 

```{r echo=FALSE}
# frq(SSC$IER_d_orig)
# frq(SSC$IER_d)

SSC %>% 
  st_drop_geometry() %>% 
  frq(IER_d_orig == IER_d)
```

```{r echo=FALSE}
plot_xtab(SSC$IER_d, SSC$IER_d_orig, 
          margin = "row", bar.pos = "stack",
          show.summary = TRUE, coord.flip = TRUE)

tabz <- table(SSC$IER_d, SSC$IER_d_orig)

assocplot(tabz,
          xlab = "IER_d", ylab = "IER_d_orig")

mosaicplot(tabz,
           xlab = "IER_d", ylab = "IER_d_orig")
```

```{r echo=FALSE}
tm_shape(SSC) +
  tm_polygons(col = c("IER_d", "IER_d_orig"),
              n = 10, palette = "div",
              id = "SSC_NAME16", 
              popup.vars = c("SSC_NAME16", "IER_d", "IER"))
```

### `IEO_d` 

```{r echo=FALSE}
# frq(SSC$IEO_d_orig)
# frq(SSC$IEO_d)

SSC %>% 
  st_drop_geometry() %>% 
  frq(IEO_d_orig == IEO_d)
```

```{r echo=FALSE}
plot_xtab(SSC$IEO_d, SSC$IEO_d_orig, 
          margin = "row", bar.pos = "stack",
          show.summary = TRUE, coord.flip = TRUE)

tabz <- table(SSC$IEO_d, SSC$IEO_d_orig)

assocplot(tabz,
          xlab = "IEO_d", ylab = "IEO_d_orig")

mosaicplot(tabz,
           xlab = "IEO_d", ylab = "IEO_d_orig")
```

```{r echo=FALSE}
tm_shape(SSC) +
  tm_polygons(col = c("IEO_d", "IEO_d_orig"),
              n = 10, palette = "div",
              id = "SSC_NAME16", 
              popup.vars = c("SSC_NAME16", "IEO_d", "IEO"))
```

```{r include=FALSE}
rm(SEIFA)
```

<!-- ------------------------------------------------------------ --> 

# Dog cost 

```{r}
dog_cost <- read_xlsx("data-raw/costs/dog_expensive.xlsx") %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>% 
  select(-web_source) %>% 
  rename(dog_breed = breed) %>%  
  select(dog_breed) %>% 
  distinct() %>% 
  # correcting names for better matching - these one used in BCE
  mutate(
    dog_breed = case_when(
      dog_breed == "Hairless Chinese Crested" ~ "Chinese Crested Dog",
      dog_breed == "Saint Bernard" ~ "St Bernard",
      TRUE ~ as.character(dog_breed))
  ) %>% 
  # synonyms
  add_row(dog_breed = "Dogue de Bordeaux") %>% 
  add_row(dog_breed = "Bulldog") %>% 
  add_row(dog_breed = "British Bulldog") %>% 
  mutate(expensive = "yes") %>% 
  arrange(dog_breed)
```

```{r include=FALSE}
write_rds(dog_cost, "data/costs/dog_cost.Rds")
```

Top 20 most expensive dogs (+3 synonyms!)

```{r echo=FALSE}
dog_cost %>% 
  datatable()
```

<!-- ------------------------------------------------------------ --> 

# Dog insurance

Data scraped from https://top10petinsurance.com.au/pet-insurance-prices on the 30th March 2020

```{r eval=FALSE}
# packages needed
# install.packages("rvest")

p_load(rvest, tidyverse)

# scraping table done with this using chrome: https://www.r-bloggers.com/using-rvest-to-scrape-an-html-table/
url <- 'https://top10petinsurance.com.au/pet-insurance-prices/'   

dog_insurance <- url %>%
  xml2::read_html() %>%
  html_nodes(xpath='//*[@id="post-1016"]/div/table') %>%
  html_table()

dog_insurance <- dog_insurance[[1]]

head(dog_insurance)

write_rds(dog_insurance, "data-raw/costs/dog_insurance.Rds") # extracted on the 30 March 2020
```

```{r}
dog_insurance <- read_rds("data-raw/costs/dog_insurance.Rds") %>% 
  as_tibble() %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>% 
  # select(-) %>% 
  mutate(average_accident_policy_cost_annual = 
           gsub(",", "",
                average_accident_policy_cost_annual, 
                fixed = TRUE),
         average_illness_policy_cost_annual = 
           gsub(",", "",
                average_illness_policy_cost_annual, 
                fixed = TRUE),
         average_comprehensive_policy_cost_annual = 
           gsub(",", "",
                average_comprehensive_policy_cost_annual, 
                fixed = TRUE)
  ) %>% 
  mutate(average_accident_policy_cost_annual = 
           as.numeric(gsub("$", "",
                           average_accident_policy_cost_annual, 
                           fixed = TRUE)),
         average_illness_policy_cost_annual = 
           as.numeric(gsub("$", "",
                           average_illness_policy_cost_annual, 
                           fixed = TRUE)),
         average_comprehensive_policy_cost_annual = 
           as.numeric(gsub("$", "",
                           average_comprehensive_policy_cost_annual, 
                           fixed = TRUE))
  ) %>% 
  filter(dog_breed != "Unknown Dog Breed")
```

```{r include=FALSE}
write_rds(dog_insurance, "data/costs/dog_insurance.Rds") 
```

## Three major categories

```{r echo=FALSE}
dog_insurance %>% 
  frq(cost_compared_to_other_breeds) 
```

## Individual breeds

Breeds in `Significantly above average` category:  

```{r echo=FALSE}
dog_insurance %>% 
  filter(cost_compared_to_other_breeds == "Significantly above average") %>% 
  select(dog_breed, average_comprehensive_policy_cost_annual) %>%  arrange(desc(average_comprehensive_policy_cost_annual)) %>% 
  datatable()
```

Breeds in `Above average` category:  

```{r echo=FALSE}
dog_insurance %>% 
  filter(cost_compared_to_other_breeds == "Above average") %>% 
  select(dog_breed, average_comprehensive_policy_cost_annual) %>% 
  arrange(desc(average_comprehensive_policy_cost_annual)) %>% 
  datatable()
```

<!-- ------------------------------------------------------------ --> 

# Dog ownership in Brisbane

```{r include=FALSE}
dog_ownership <- read_csv("data-raw/permits/cars-bis-open-data-animal-permits-3-jan-2019.zip")
```

Raw dataset named `cars-bis-open-data-animal-permits-3-jan-2019` consists of `r comma(nrow(dog_ownership))` records.  

## Data selection 

**Excluding** records with:  

- `permit_name`: *Breeders Permit*, *Cattery Permit*, *Racehorses Permit*, *Pet Shop Permit*, *Domestic Dog Permit* & *Guard Dog Permit*  
- without neighbourhood  
- without `dog_breed` values  
- with `dog_breed` listed as *Unknown* or *Cross*  

*Stones Corner* values were assigned to *Greenslopes* SSC (see information in section above).   

```{r}
dog_ownership %<>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>% 
  # all the same here
  select(-permit_group, -permit_status) %>% 
  # special permits?
  filter(!permit_name %in% c("Breeders Permit",	"Cattery Permit", "Racehorses Permit", "Pet Shop Permit")) %>% 
  filter(!permit_name %in% c("Domestic Dog Permit", "Guard Dog Permit")) %>%
  rename(dog_breed = animal_breed,
         SSC_NAME16 = application_location_suburb) %>% 
  mutate(SSC_NAME16 = str_to_title(SSC_NAME16)) %>% 
  # correct suburb
  mutate(SSC_NAME16 = ifelse(SSC_NAME16 == "Stones Corner", "Greenslopes", SSC_NAME16)) %>% 
  # missing geo
  filter(!is.na(SSC_NAME16)) %>% 
  # missing breed
  filter(!is.na(dog_breed)) %>% 
  filter(!dog_breed %in% c("Unknown", 
                           "Medium Cross Breed", "Small Cross Breed", "Large Cross Breed")) %>% 
  # few cleans for better matches
  mutate(
    dog_breed = case_when(
      dog_breed == "German Shepherd Dog (Long Stock Coat)" ~ "German Shepherd",
      dog_breed == "German Shepherd Dog" ~ "German Shepherd",
      dog_breed == "Central Asian Shepherd Dog " ~ "Central Asian Shepherd",
      dog_breed == "Kangal Dog" ~ "Kangal",
      dog_breed == "Bulldog" ~ "British Bulldog",
      dog_breed == "Collie (Rough)" ~ "Rough Collie", 
      dog_breed == "Collie (Smooth)" ~ "Smooth Collie",      
      
      TRUE ~ as.character(dog_breed))
  ) 
```

```{r include=FALSE}
write_rds(dog_ownership, "data/permits/dog_ownership.Rds") 
```

Prepared dataset consists of `r comma(nrow(dog_ownership))` records.  

```{r eval=FALSE, include=FALSE}
# checking domestic dogs
dog_ownership %>% 
  filter(permit_name %in% c("Domestic Dog Permit")) %>%
  frq(SSC_NAME16)

test <- dog_ownership %>% 
  mutate(ind = ifelse(permit_name == "Domestic Dog Permit", 1, 0)) %>% 
  group_by(SSC_NAME16) %>% 
  summarise(ind = sum(ind), total = n()) %>% 
  mutate(perc = (ind/total)*100)
```

## Data merging - `dog_cost`

```{r}
dog_ownership_cost <- left_join(dog_ownership, dog_cost, by = "dog_breed") %>% 
  # binary indicator "expensive" and "non-expensive" dog breeds according to dog_cost
  mutate(
    expensive = case_when(
      is.na(expensive) ~ 0,
      expensive == "yes" ~ 1
    ),
    expensive = as.integer(expensive)
  )
```

Expensive dogs in Brisbane:  

```{r echo=FALSE}
frq(dog_ownership_cost, expensive)
```

## Data merging - `dog_insurance`

Correcting names for better match

```{r}
dog_insurance %<>%
  mutate(
    dog_breed = case_when(
      # this is a bit tricky! might need sensitivity?
      dog_breed == "French Poodle" ~ "Poodle",  
      
      # just naming issues
      dog_breed == "Poodle – Standard" ~ "Poodle (Standard)",
      dog_breed == "Miniature Poodle" ~ "Poodle (Miniature)",
      dog_breed == "Poodle – Toy" ~ "Poodle (Toy)",
      dog_breed == "Shar-Pei" ~ "Shar Pei", 
      dog_breed == "German Short Haired Pointer" ~ "German Shorthaired Pointer", 
      dog_breed == "German Wire Haired Pointer" ~ "German Wirehaired Pointer", 
      dog_breed == "Collie – Rough" ~ "Rough Collie", 
      dog_breed == "Collie – Smooth" ~ "Smooth Collie", 
      dog_breed == "Miniature Schnauzer" ~ "Schnauzer (Miniature)", 
      dog_breed == "Schnauzer Giant" ~ "Schnauzer (Giant)", 
      dog_breed == "Lagotto Rom" ~ "Lagotto Romagnolo", 
      dog_breed == "Brittany Spaniel" ~ "Brittany", 
      dog_breed == "Staghound" ~ "Stag Hound", 
      dog_breed == "Kerry Blue" ~ "Kerry Blue Terrier", 
      dog_breed == "English Toy terrier" ~ "English Toy Terrier",
      dog_breed == "Parson Jack Russell Terrier" ~ "Parson Russell Terrier", 
      dog_breed == "Welsh Corgi – Pembroke" ~ "Welsh Corgi (Pembroke)", 
      dog_breed == "American Cocker Spaniel" ~ "Cocker Spaniel (American)",
      dog_breed == "Basset Fauve De Bretagne" ~ "Basset Fauve de Bretagne", 
      dog_breed == "Norwegian Elk Hound" ~ "Norwegian Elkhound", 
      dog_breed == "Cheasapeake Bay Retriever" ~ "Chesapeake Bay Retriever", 
      dog_breed == "Bouvier Des Flandres" ~ "Bouvier des Flandres", 
      dog_breed == "Miniature Bull Terrier" ~ "Bull Terrier (Miniature)", 
      dog_breed == "Munsterlander – Large" ~ "Large Munsterlander", 
      dog_breed == "Welsh Corgi – Cardigan" ~ "Welsh Corgi (Cardigan)", 
      dog_breed == "HamiltonStovare" ~ "Hamiltonstovare",
      dog_breed == "Blue tick Coonhound" ~ "Bluetick Coonhound",
      dog_breed == "Japanese Akita" ~ "Akita (Japanese)",
      dog_breed == "Dogue De Bordeaux" ~ "Dogue de Bordeaux",
      dog_breed == "Italian Cane Corso" ~ "Italian Corso Dog",
      
      # different kelpies but same category anyway
      dog_breed == "Australian Kelpie Sheepdog" ~ "Australian Kelpie", 
      
      # typo
      dog_breed == "Neopolitan Mastiff" ~ "Neapolitan Mastiff", 
      TRUE ~ as.character(dog_breed)))
```

Dog breeds without match 

```{r echo=FALSE}
dog_ownership_cost %>% anti_join(
  dog_insurance %>% select(dog_breed, cost_compared_to_other_breeds)
) %>%
  group_by(dog_breed) %>%
  summarize(n = n()) %>%
  arrange(desc(n)) %>% 
  datatable() 
```

Some further corrections still possible here:

```{r}
dog_ownership_cost <-  left_join(dog_ownership_cost, 
                                 dog_insurance %>% 
                                   select(dog_breed, cost_compared_to_other_breeds)
) %>% 
  mutate(
    cost_compared_to_other_breeds = case_when(
      # all the same
      dog_breed == "Fox Terrier" ~ "Below average",
      dog_breed == "Fox Terrier (Smooth)" ~ "Below average",
      dog_breed == "Fox Terrier (Wire)" ~ "Below average",
      dog_breed == "Schnauzer" ~ "Below average",
      dog_breed == "Australian Stumpy Tail Cattle Dog" ~ "Below average",
      dog_breed == "Foxhound" ~ "Below average",
      dog_breed == "White Swiss Shepherd Dog" ~ "Below average",
      
      # multiple options here, but all above going for conservative
      dog_breed == "Welsh Corgi" ~ "Above average",
      
      # taking values from Chihuahua
      dog_breed == "Chihuahua (Smooth Coat)" ~ "Below average",
      dog_breed == "Chihuahua (Long Coat)" ~ "Below average",
      dog_breed == "Chihuahua (Long Coat)" ~ "Below average",
      
      # naming - Dogue De Bordeaux
      dog_breed == "French Mastiff" ~ "Significantly above average",
      TRUE ~ as.character(cost_compared_to_other_breeds))) %>% 
  select(-permit_name)
```

Few things left:  

```{r echo=FALSE}
dog_ownership_cost %>% 
  filter(is.na(cost_compared_to_other_breeds)) %>%
  group_by(dog_breed) %>%
  summarize(n = n()) %>%
  arrange(desc(n)) %>% 
  datatable() 
```

Two largest groups:  

- *Pointer* - insufficient info!  
- *Akita* - insufficient info! Could be Inu, could be Japanese  

These observations will remain as `NA`s.  

```{r include=FALSE}
rm(dog_ownership, dog_cost, dog_insurance)
```

<!-- ----------------------------------------------------- -->

# Dogs per capita

Summarizing all dogs, and expensive only.  

```{r}
dog_agg <- dog_ownership_cost %>% 
  group_by(SSC_NAME16) %>% 
  summarise(dogs_total = n())

SSC %<>% 
  left_join(dog_agg)
```

## Missing SEIFA 

Areas with no URP/SEIFA but having (small amount of) dogs:

```{r echo=FALSE}
temp1 <- SSC %>% 
  st_drop_geometry() %>% 
  right_join(dog_agg) %>% 
  filter(is.na(SSC_CODE16)) %>% 
  select(SSC_NAME16, dogs_total)

temp1

dog_ownership_cost %<>% 
  filter(!SSC_NAME16 %in% temp1$SSC_NAME16)
```

These areas were excluded from the `dog_ownership_cost` dataset.  

## Missing dogs 

Area with low URP (also marked as caution == TRUE in SEIFA) and no dogs at all:

```{r echo=FALSE}
temp2 <- SSC %>% 
  st_drop_geometry() %>% 
  left_join(dog_agg) %>% 
  filter(is.na(dogs_total)) %>% 
  select(SSC_NAME16, dogs_total, URP)

temp2

SSC %<>% 
  filter(!SSC_NAME16 %in% temp2$SSC_NAME16)
```

```{r}
SSC %<>% 
  filter(dogs_total > 0) %>% 
  mutate(dogs_percap = (dogs_total / URP)*100) %>% 
  relocate(geometry, .after = last_col())
```

This area was excluded.  

## Ranking

```{r echo=FALSE}
SSC %>% 
  st_drop_geometry() %>% 
  arrange(desc(dogs_percap)) %>% 
  select(SSC_NAME16, dogs_total, URP, dogs_percap) %>% 
  mutate(dogs_percap = number(dogs_percap, accuracy = 0.1)) %>% 
  # slice(1:10) %>% 
  datatable()
```

## Map

```{r echo=FALSE}
tm_shape(SSC) +
  tm_bubbles(id = "SSC_NAME16", 
             size = "dogs_total", 
             col = "dogs_percap",
             style = "quantile",
             palette = "BrBG",
             border.col = NA)
```

```{r include=FALSE}
write_rds(dog_ownership_cost, "data/dog_ownership_cost.Rds")
write_rds(SSC, "data/geo/SSC.Rds")
```

<!-- ----------------------------------------------------- -->

# Wide format datasets

## All breeds

```{r}
wide_all_n <- dog_ownership_cost %>% 
  group_by(SSC_NAME16, dog_breed) %>% 
  summarise(n = n()) %>% 
  ungroup() %>% 
  pivot_wider(names_from = dog_breed, 
              values_from = n) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  clean_names() %>% 
  rename(SSC_NAME16 = ssc_name16)

wide_all_n <- right_join(SSC, wide_all_n)

wide_all_p <- dog_ownership_cost %>% 
  group_by(SSC_NAME16, dog_breed) %>% 
  summarise(n = n()) %>% 
  ungroup() %>% 
  group_by(SSC_NAME16) %>% 
  mutate(N = sum(n)) %>% 
  ungroup() %>% 
  mutate(p = n/N) %>% 
  select(-n, -N) %>% 
  pivot_wider(names_from = dog_breed, 
              values_from = p) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  clean_names() %>% 
  rename(SSC_NAME16 = ssc_name16)

wide_all_p <- right_join(SSC, wide_all_p)
```

## Only expensive breeds (cost)

```{r}
wide_cost_n <- dog_ownership_cost %>% 
  filter(expensive == 1) %>% 
  group_by(SSC_NAME16, dog_breed) %>% 
  summarise(n = n()) %>% 
  ungroup() %>% 
  pivot_wider(names_from = dog_breed, 
              values_from = n) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  clean_names() %>% 
  rename(SSC_NAME16 = ssc_name16)

wide_cost_n <- right_join(SSC, wide_cost_n)

wide_cost_p <- dog_ownership_cost %>% 
  filter(expensive == 1) %>% 
  group_by(SSC_NAME16, dog_breed) %>% 
  summarise(n = n()) %>% 
  ungroup() %>% 
  group_by(SSC_NAME16) %>% 
  mutate(N = sum(n)) %>% 
  ungroup() %>% 
  mutate(p = n/N) %>% 
  select(-n, -N) %>% 
  pivot_wider(names_from = dog_breed, 
              values_from = p) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  clean_names() %>% 
  rename(SSC_NAME16 = ssc_name16)

wide_cost_p <- right_join(SSC, wide_cost_p)
```

## Only expensive breeds (insurance)

```{r}
wide_insurance_n <- dog_ownership_cost %>% 
  # filter(cost_compared_to_other_breeds != "Below average") %>% 
  filter(cost_compared_to_other_breeds == "Significantly above average") %>% 
  group_by(SSC_NAME16, dog_breed) %>% 
  summarise(n = n()) %>% 
  ungroup() %>% 
  pivot_wider(names_from = dog_breed, 
              values_from = n) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  clean_names() %>% 
  rename(SSC_NAME16 = ssc_name16)

wide_insurance_n <- right_join(SSC, wide_insurance_n)

wide_insurance_p <- dog_ownership_cost %>% 
  # filter(cost_compared_to_other_breeds != "Below average") %>% 
  filter(cost_compared_to_other_breeds == "Significantly above average") %>% 
  group_by(SSC_NAME16, dog_breed) %>% 
  summarise(n = n()) %>% 
  ungroup() %>% 
  group_by(SSC_NAME16) %>% 
  mutate(N = sum(n)) %>% 
  ungroup() %>% 
  mutate(p = n/N) %>% 
  select(-n, -N) %>% 
  pivot_wider(names_from = dog_breed, 
              values_from = p) %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  clean_names() %>% 
  rename(SSC_NAME16 = ssc_name16)

wide_insurance_p <- right_join(SSC, wide_insurance_p)
```

```{r include=FALSE}
write_rds(wide_all_n, "data/wide_all_n.Rds")
write_rds(wide_all_p, "data/wide_all_p.Rds")
write_rds(wide_cost_n, "data/wide_cost_n.Rds")
write_rds(wide_cost_p, "data/wide_cost_p.Rds")
write_rds(wide_insurance_n, "data/wide_insurance_n.Rds")
write_rds(wide_insurance_p, "data/wide_insurance_p.Rds")
```

<!-- ------------------------------------------------------------ --> 

# ML data 

## Saves 

Using the Index of Education and Occupation (`IEO`) only as outcome. More info about the index [from ABS](https://www.abs.gov.au/ausstats/abs@.nsf/Lookup/by%20Subject/2033.0.55.001~2016~Main%20Features~IEO~22).  

Keeping all breeds and wide formats as described above; droppping spatial info.  

```{r}
wide_all_n %<>% 
  sf::st_drop_geometry() %>% 
  # as_tibble() %>% 
  select(IEO, IEO_d, 
         akita:last_col())

wide_all_p %<>%  
  sf::st_drop_geometry() %>% 
  # as_tibble() %>% 
  select(IEO, IEO_d, 
         akita:last_col()) %>% 
  rename(outcome = IEO,
         outcome_d = IEO_d)
```

```{r eval=FALSE, include=FALSE}
tail(names(wide_all_p), n = 1)
skimr::skim(wide_all_p)
report::report(wide_all_p)
```

## Near zero-variance 

```{r}
nzv <- nearZeroVar(wide_all_p, saveMetrics = TRUE)
```

```{r echo=FALSE}
nzv %>% 
  filter(nzv == TRUE) %>% 
  datatable()
```

```{r}
nzv <- nearZeroVar(wide_all_p)
wide_all_p <- wide_all_p[, -nzv]
wide_all_n <- wide_all_n[, -nzv]
rm(nzv)
```

```{r}
wide_all_n %<>% 
  mutate(weight = rowSums(across(where(is.numeric)))) %>% 
  select(weight)
```

## Correlated breeds

```{r}
descr_cor <- cor(wide_all_p %>% select(-outcome, -outcome_d)) %>% 
  as_tibble(rownames = "rowname") %>% 
  pivot_longer(-rowname) %>% 
  filter(rowname != name) %>% 
  arrange(desc(value)) %>% 
  filter(row_number() %% 2 == 1) %>% 
  mutate(value = scales::number(value, accuracy = 0.01)) %>% 
  rename(correlation = value)
```

```{r echo=FALSE}
descr_cor %>% 
  datatable()
```

### Linear combinations

```{r}
combo_info <- findLinearCombos(wide_all_p %>% select(-outcome, -outcome_d))
combo_info
```

## Train split

75/25 split, taking into account distribution of deciles:    

```{r}
inTrain <- createDataPartition(
  y = wide_all_p$outcome_d,
  p = .75,
  list = FALSE
)

training <- wide_all_p[ inTrain, ]
testing  <- wide_all_p[-inTrain, ]

training_weights <- wide_all_n[ inTrain,]
```

Training data:  

```{r echo=FALSE}
# nrow(training)
frq(training, outcome_d)
```

Testing data:  

```{r echo=FALSE}
# nrow(testing)
frq(testing, outcome_d)
```

```{r include=FALSE}
training <- select(training, -outcome_d)
testing <- select(testing, -outcome_d)
```

Outcome distro checks  

```{r echo=FALSE}
ggplot() +
  geom_density(data = training, aes(x = outcome), col = "forestgreen") +
  geom_density(data = testing, aes(x = outcome), col = "firebrick") 
```

```{r include=FALSE}
write_rds(training, "data/training.Rds")
write_rds(testing, "data/testing.Rds")
write_rds(training_weights, "data/training_weights.Rds")
```

<!-- ----------------------------------------------------- -->

# Computing Environment

`r mu$session()`
