==> R/06_nn.Rmd <==
---
title: "Dog-SEP"
subtitle: "Machine learning: neural networks"
pagetitle: "Dog-SEP: NN"
author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document: 
    highlight: pygments
    keep_md: no
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 4
    self_contained: true
    toc_float: yes
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ------------------------------------------------------------ --> 

```{r r-setup, include=FALSE}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")

library(pacman)
p_load(tidyverse, magrittr, DT,
       caret, neuralnet)

import::from("sjmisc", "frq")
```

```{r knit-setup, include=FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, 
                      fig.width = 8, fig.height = 6, dpi = 300, 
                      out.width = "800px", out.height = "600px")

knitr::opts_chunk$set(cache = FALSE,
                      prompt = FALSE,
                      tidy = FALSE,
                      comment = NA,
                      message = FALSE,
                      warning = FALSE)

knitr::opts_knit$set(width = 75)

mu <- Hmisc::markupSpecs$html
```

```{r include=FALSE}
# function to calculate the mean absolute error
MAE <- function(actual, predicted) {
  mean(abs(actual - predicted))  
}
```

<!-- ------------------------------------------------------------ --> 

# Background 

Using info from chapter 7 of *Machine Learning with R* [book](https://www.packtpub.com/product/machine-learning-with-r-third-edition/9781788295864), `caret` package [manual](https://topepo.github.io/caret/index.html), and `neuralnet` package [vignette](https://cran.r-project.org/web/packages/neuralnet/).  

<!-- ------------------------------------------------------------ --> 

# Analysis with `neuralnet`

## Data 

```{r include=FALSE}
wide_all_p <- read_rds("data/wide_all_p.Rds")

wide_all_p %<>%  
  sf::st_drop_geometry() %>% 
  # as_tibble() %>% 
  select(IEO, IEO_d, 
         akita:last_col()) %>% 
  rename(outcome = IEO)

nzv <- nearZeroVar(wide_all_p, saveMetrics = TRUE)
nzv <- nearZeroVar(wide_all_p)
wide_all_p <- wide_all_p[, -nzv]

inTrain <- createDataPartition(
  y = wide_all_p$IEO_d,
  p = .75,
  list = FALSE
)

wide_all_p <- select(wide_all_p, -IEO_d)

normalize <- function(x) {
  return((x - min(x)) / (max(x) - min(x)))
}

wide_all_p <- as.data.frame(lapply(wide_all_p, normalize))

training <- wide_all_p[ inTrain, ]
testing  <- wide_all_p[-inTrain, ]
```

## Training

```{r}
set.seed(12345)
m_nn <- neuralnet(formula = outcome ~ ., 
                  data = training, 
                  hidden = c(5, 5))
```

```{r echo=FALSE}
summary(m_nn)
```

## Prediction

```{r}
p_nn <- predict(m_nn, testing[-1])
```

Correlation:  

```{r}
cor(p_nn, testing$outcome)
```

MAE between predicted and actual values:  

```{r}
MAE(p_nn, testing$outcome)
```

Distro of predicted (red) nd actual (green): 

```{r echo=FALSE}
# summary(p_nn)
# summary(testing$outcome)

ggplot() +
  geom_density(data = as.data.frame(p_nn), aes(x = p_nn), col = "firebrick") +
  geom_density(data = testing, aes(x = outcome), col = "forestgreen") 
```

## Tuning

```{r}
getModelInfo("neuralnet")
```

```{r}
modelLookup("neuralnet")
```

Using grid of parameters to explore:  

```{r}
grid_tu <-  expand.grid(layer1  = seq(15, 25, 1),
                        layer2  = seq(15, 25, 1),
                        layer3  = (1))

nrow(grid_tu)
```

```{r}
p_load(doParallel)

cl <- makePSOCKcluster(parallel::detectCores())
registerDoParallel(cl)

set.seed(12345)
m_nn_tu <- train(outcome ~ .,
                 method = "neuralnet",
                 data = training,
                 # weights = training_weights, 
                 metric = "RMSE",
                 # metric = "Rsquared",
                 tuneGrid = grid_tu)

stopCluster(cl)
p_unload(doParallel)
```

```{r echo=FALSE}
# summary(m_nn_tu)
ggplot(m_nn_tu)
```

```{r}
m_nn_tu$finalModel
```

```{r}
p_nn_tu <- predict(m_nn_tu, testing[-1])
```

Correlation:  

```{r}
cor(p_nn_tu, testing$outcome)
```

MAE between predicted and actual values:  

```{r}
MAE(p_nn_tu, testing$outcome)
```

Distro of predicted (red) nd actual (green): 

```{r echo=FALSE}
# summary(p_nn)
# summary(testing$outcome)

ggplot() +
  geom_density(data = as.data.frame(p_nn_tu), aes(x = p_nn_tu), col = "firebrick") +
  geom_density(data = testing, aes(x = outcome), col = "forestgreen") 
```
