==> R/12_h2o.Rmd <==
---
title: "Dog-SEP"
subtitle: "Machine learning: h2o & DALEX"
pagetitle: "Dog-SEP: h2o"
author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document: 
    highlight: pygments
    keep_md: no
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 4
    self_contained: true
    toc_float: yes
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<!-- ------------------------------------------------------------ --> 

```{r r-setup}
set.seed(12345)
options(scipen = 999)
options(max.print = "75")

library(pacman)
p_load(
  tidyverse, magrittr, DT,
  h2o, 
  DALEX, DALEXtra
)
```

```{r knit-setup, include=FALSE}
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file(),
                     width = 75)

knitr::opts_chunk$set(
  echo = TRUE,
  cache = FALSE,
  prompt = FALSE,
  tidy = FALSE,
  comment = NA,
  message = FALSE,
  warning = FALSE,
  fig.width = 8, fig.height = 6, dpi = 300,
  out.width = "800px", out.height = "600px"
)
```

```{r eval=FALSE}
if ("package:h2o" %in% search()) {
  detach("package:h2o", unload = TRUE)
}
if ("h2o" %in% rownames(installed.packages())) {
  remove.packages("h2o")
}

p_load("RCurl", "jsonlite")

install.packages("h2o", type = "source", repos = (c("http://h2o-release.s3.amazonaws.com/h2o/latest_stable_R")))
```

<!-- ------------------------------------------------------------ --> 

# Data 

```{r include=FALSE}
training <- read_rds("data/training.Rds")
testing <- read_rds("data/testing.Rds")

training_weights <- read_rds("data/training_weights.Rds")

training <- training %>% 
  bind_cols(training_weights) %>% 
  rename( weights = `...122`) %>% 
  mutate(weights = (weights-min(weights)) / (max(weights)-min(weights)))
```

# Analysis with `h2o`

## Data formatting 

```{r}
h2o.init()

training <- as.h2o(training)
testing <- as.h2o(testing)
# training_weights <- as.h2o(training_weights)
```

## Model  

```{r}
m_h2o_aml <- h2o.automl(y = "outcome", 
                        training_frame = training, 
                        # weights_column = "weights", 
                        max_models = 50)
```

```{r}
h2o.get_leaderboard(m_h2o_aml) %>% 
  as.data.frame() %>%
  dplyr::select(model_id, rmse) %>%
  dplyr::slice(1:10)

m_h2o_aml@leader
```

## Predict 

```{r}
explainer <- DALEXtra::explain_h2o(m_h2o_aml, 
                                   label = "h2o", 
                                   data = testing, y = testing$outcome, 
                                   verbose = FALSE)
```

```{r eval=FALSE}
new_obs <- testing[1, ]

pb <- predict_parts(explainer, 
                    new_observation = new_obs, 
                    type = "break_down")

plot(pb)
```

```{r eval=FALSE, include=FALSE}
# `modelStudio` 
library(modelStudio)

# set up the cluster
options(
  parallelMap.default.mode        = "socket",
  parallelMap.default.cpus        = 9,
  parallelMap.default.show.info   = FALSE
)

# calculations of local explanations will be distributed into 4 cores
modelStudio(explainer,
            new_observation = testing[1:10,],
            parallel = TRUE)
```

```{r}
model_path <- h2o.saveModel(object = m_h2o_aml@leader, 
                            path = "results", force = TRUE,
                            filename = "m_h2o_aml"
                            # filename = "m_h2o_aml_weight"
                            )
print(model_path)
```
